Merge `client` feature/module into `gateway` (#2947)
The `client` and `gateway` features were very tightly coupled so it
makes the most sense to just merge them because there's no real path
towards separating them ergonomically.

A few breaking changes:
* Removed `ClientError` because it was unused.
* The `client` module is moved to be a submodule under the `gateway`
module.
* Moved sharding-specific gateway types into their own `sharding`
submodule for clarity's sake, but re-exported them.

I also adjusted the gateway documentation to hopefully make it as clear as
possible.

Note that the `Cargo.toml` changes are mostly superficial - in
particular, changing `gateway` to depend on `model` (instead of `http`
like how `client` did) isn't actually much of a change because trying to
use `gateway` with only a subset of `model` enabled will fail to
compile, and that was also the case for `client` previously.