Fix `ShardQueue` batch size bug (#3121)
If the last batch to be started was not of maximum size, it wouldn't get popped
off the queue. Instead, we should just pop off the queue periodically to ensure
that we fully flush it. 

On client startup, we queue all shards before popping the first batch, so
suboptimal batch sizes are not a concern. Now, the queue also doesn't have to
worry about "concurrent mode" and saturating the size of its batches.